Fix issue where ad type creation was broken
The ApiEndpoint class takes only :client and :endpoint constructor
arguments. It then uses the endpoint in both the URL it POSTs, GETs,
and PUTs to to create, view, and update records.

Consider the sites endpoint [1], for instance:

Create new site:

    POST /v1/site HTTP/1.1
    Content-Type: application/x-www-form-urlencoded

    site={...}

List all sites:

    GET /v1/site HTTP/1.1

Update site:

    PUT /v1/site/123 HTTP/1.1
    Content-Type: application/x-www-form-urlencoded

    site={...}

However, for the adtypes endpoint [2] we have a problem:

Create new ad type:

    POST /v1/adtypes HTTP/1.1
    Content-Type: application/x-www-form-urlencoded

    adtype={...}

Notice how the key in the urlencoded body (adtype) is different from
the endpoint (adtypes).

This commit adds a :datakey parameter to the ApiEndpoint class
constructor arguments to accomodate this special case. This new
parameter is optional and defaults to the value of :endpoint.

[1]: https://github.com/adzerk/adzerk-api/wiki/Site-Endpoints
[2]: https://github.com/adzerk/adzerk-api/wiki/Ad-Type-Endpoints