Add optional channel param to create, list, and delete methods
This commit adds an optional parameter, the channel id, to the
ApiEndpoint class' @create, @list, and @delete methods. This parameter
was added to support the idiosyncracies of the ad types endpoint [1]
where the API has been split into operations that are scoped to a
particular channel and operations that are applied to the entire
network.

Use the extra channel parameter when creating or deleting ad types for a
specific channel, and when listing the ad types for a channel.