temporary fix to work around the error message being about to change
The route lockdown in progress for API v1 will start throwing a 302 "Object moved" HTML response for these endpoints when a user tries to access/modify ads in a flight not owned by his/her network.

The long-term solution is API v2. For now, I think it's sufficient that trying to access/modify things you don't own results in an error (albeit not an informative one), since this should not be part of a typical Adzerk API workflow.