fix Gemfile vs. adzerk.gemspec conflicts; use rest-client 2.0.1 for real
The root of the problem here was that there were dependencies specified in both
the Gemfile and adzerk.gemspec.

A second problem is that we have Gemfile.lock checked into the repo. It is
recommended that libraries (such as this one) should not have Gemfile.lock
checked into the repo.

This commit makes adzerk.gemspec the source of truth for dependencies and
Gemfile is now just a shim that points to adzerk.gemspec for development and
runtime dependency information.

Also, Gemfile.lock is now in .gitignore.