fix warning about Content-Type header actually being application/x-www-form-urlencoded
The "Create Creative" API endpoint expects application/x-www-form-urlencoded,
not application/json. Fortunately, the Ruby gem has been correctly using
x-www-form-urlencoded despite the params including `:content_type => :json`,
because the rest-client library uses x-www-form-urlencoded by default unless you
call `.to_json` on the body of the entire request.

(Note that there is a `.to_json` in the code here, but what we're actually
sending is what the API expects, which is x-www-form-urlencoded data where the
field "creative" is that JSON string.)

It generates a warning though because we're telling it to send JSON and it's
sending x-www-form-urlencoded data. This commit fixes the warning by not telling
it that we're sending JSON, because we're not.